リファクタとバグの修正

+++ src/extra/shader_node/picture.js
@@ -10,20 +10,26 @@ export const PictureNode = class extends FrameNode {
 		this.x = x;
 		this.y = y;
 		this.json["custom"].img_url = img_url;
+		this.loading = true;
 	}
-	setup(){
-		super.setup();
-		let img = this.graphics.createTexture(0, 0);
-		img.loadImg(this.json["custom"].img_url, () => {
-			this.resizeFrame(img.width, img.height);
-			this.frameBuffer.beginDraw();
-			this.graphics.image(img, 0, 0, img.width, img.height);
-			this.frameBuffer.endDraw();
-			this.resizeBox.target.y = this.w * this.frameBuffer.height / this.frameBuffer.width;
-		});
-		this.inputs.remove(this.inputShaderNodeParam);
-		this.inputs.remove(this.inputResolutionNodeParam);
-		img.delete();
+	update() {
+		super.update();
+		if (this.loading) {
+			this.loading = false;
+			let img = this.graphics.createTexture(0, 0);
+			img.loadImg(this.json["custom"].img_url, () => {
+				this.resizeFrame(img.width, img.height);
+				this.frameBuffer.beginDraw();
+				this.graphics.image(img, 0, 0, img.width, img.height);
+				this.frameBuffer.endDraw();
+				this.resizeBox.target.y = this.w * this.frameBuffer.height / this.frameBuffer.width;
+			}, () => {
+				this.parent.remove(this);
+			});
+			this.inputs.remove(this.inputShaderNodeParam);
+			this.inputs.remove(this.inputResolutionNodeParam);
+			img.delete();
+		}
 	}
 	job() {
 		super.job();
